main.py
from dotenv import load_dotenv

load_dotenv()

from langchain_classic import hub
from langchain_classic.agents import AgentExecutor, create_react_agent
from langchain_openai import ChatOpenAI
from langchain_tavily import TavilySearch



def main():
    print("Hello from pythonproject2!")


if __name__ == "__main__":
    main()
